알림 불러오는 로직을 custom hook으로 추출하기
